refactor databricksclient
