+++ source/dotnet/Wholesale.sln
@@ -20,8 +20,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ConfigurationFiles", "Confi
 		stylecop.json = stylecop.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DatabricksClient", "wholesale-api\DatabricksClient\DatabricksClient.csproj", "{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApi.IntegrationTests", "wholesale-api\WebApi.IntegrationTests\WebApi.IntegrationTests.csproj", "{6D00C8EA-65D9-4568-B3C6-23E31598D51E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test.Core", "wholesale-api\Test.Core\Test.Core.csproj", "{EFA0CC43-7B22-4842-9AC0-E0FDA4F844BA}"
@@ -92,10 +90,6 @@ Global
 		{29B77280-062B-4AB0-A67D-8AF9C346299F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{29B77280-062B-4AB0-A67D-8AF9C346299F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{29B77280-062B-4AB0-A67D-8AF9C346299F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6D00C8EA-65D9-4568-B3C6-23E31598D51E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6D00C8EA-65D9-4568-B3C6-23E31598D51E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D00C8EA-65D9-4568-B3C6-23E31598D51E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -174,7 +168,6 @@ Global
 		{EFA0CC43-7B22-4842-9AC0-E0FDA4F844BA} = {A65AE4F0-E4A5-4F6C-AB74-F8A278C31AD9}
 		{F163CC59-E435-4274-9316-7EF7463BA8E6} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
 		{A65AE4F0-E4A5-4F6C-AB74-F8A278C31AD9} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
-		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
 		{9746F437-9232-4F33-B11C-DBE78AF88A79} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
 		{B1CB77BA-0D18-410C-BC65-DDA21275E99E} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
 		{AEB36AF9-5E4E-4BF9-990D-F6D65E9E2482} = {B64D5473-F242-4D51-80AA-46DE0CF39C8A}
