+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Batches.Infrastructure.csproj
@@ -8,7 +8,6 @@
     <PackageReference Include="NodaTime" Version="3.1.9" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\DatabricksClient\DatabricksClient.csproj" />
     <ProjectReference Include="..\Batches.Application\Batches.Application.csproj" />
   </ItemGroup>
 </Project>
