+++ source/dotnet/wholesale-api/Batches/Batches.UnitTests/Infrastructure/Calculations/DatabricksCalculatorJobRunnerTests.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Wholesale.Batches.Application;
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.Calculations;
-using Energinet.DataHub.Wholesale.Components.DatabricksClient.DatabricksWheelClient;
+using Energinet.DataHub.Wholesale.Common.DatabricksClient;
 using Microsoft.Azure.Databricks.Client;
 using Microsoft.Azure.Databricks.Client.Models;
 using Moq;
@@ -48,7 +48,7 @@ public class DatabricksCalculatorJobRunnerTests
         CalculationState expectedCalculationState,
         RunLifeCycleState runLifeCycleState,
         RunResultState runResultState,
-        [Frozen] Mock<IDatabricksWheelClient> databricksWheelClientMock,
+        [Frozen] Mock<IJobsApiClient> databricksWheelClientMock,
         CalculationEngineClient sut)
     {
         var jobRunId = new CalculationId(1);
