+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.Infrastructure/CalculationResults.Infrastructure.csproj
@@ -11,7 +11,6 @@
     <PackageReference Include="NodaTime" Version="3.1.9" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\DatabricksClient\DatabricksClient.csproj" />
     <ProjectReference Include="..\CalculationResults.Application\CalculationResults.Application.csproj" />
   </ItemGroup>
 </Project>
