+++ source/dotnet/wholesale-api/Common/DatabricksClient/DatabricksOptions.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Components.DatabricksClient;
+namespace Energinet.DataHub.Wholesale.Common.DatabricksClient;
 
 public class DatabricksOptions
 {
